#5 - JpaRepository 타입 수정
IDE 자동완성으로 인해 Entity id 타입을 모두 Integer로 설정하여 해당 타입에 맞게 수정함.

+++ src/main/java/com/back/repository/CommentRepository.java
@@ -3,5 +3,5 @@ package com.back.repository;
 import com.back.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
+public interface CommentRepository extends JpaRepository<Comment, Long> {
 }
